EOS-16208 Changes for Reset and Refresh Config Changes
