+++ csm/core/blogic/const.py
@@ -109,6 +109,7 @@ CSM_GLOBAL_INDEX = 'CSM'
 INVENTORY_INDEX = 'INVENTORY'
 COMPONENTS_INDEX = 'COMPONENTS'
 DATABASE_INDEX = 'DATABASE'
+CORTXCLI_GLOBAL_INDEX = 'CORTXCLI'
 
 # AMQP Consumer Tag
 CONSUMER_TAG = 'AMQP_CONSUMER'
